refactor : config 폴더의 내용들 폴더별로 정리
common : 공통 응답 형식
exception : 예외처리 관련
security : 시큐리티 설정 관련
로 정리함
