+++ src/main/java/com/web/SearchWeb/config/security/CustomAuthenticationSuccessHandler.java
@@ -1,4 +1,4 @@
-package com.web.SearchWeb.config;
+package com.web.SearchWeb.config.security;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
